feat: 🎸 Adding a new component for sidebar menu
